ML/Support_Vector_Machine/Project_Image_Classification/requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Image processing
opencv-python>=4.5.0
Pillow>=8.0.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Model persistence
joblib>=1.1.0

# Configuration
PyYAML>=6.0

# Azure ML
azureml-core>=1.40.0
azureml-automl-core>=1.40.0

# Logging
python-dotenv>=0.19.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0
